+++ hypervisor/Makefile
@@ -215,6 +215,8 @@ HW_C_SRCS += arch/x86/page.c
 HW_C_SRCS += arch/x86/notify.c
 HW_C_SRCS += arch/x86/vtd.c
 HW_C_SRCS += arch/x86/gdt.c
+HW_C_SRCS += arch/x86/nmi.c
+HW_C_SRCS += arch/x86/exception.c
 HW_C_SRCS += arch/x86/irq.c
 HW_C_SRCS += arch/x86/timer.c
 HW_C_SRCS += arch/x86/vmx.c
